refactor(node): remove pass-through wrappers and fold component-builder crates
- Replace the TaikoAddOns delegation wrapper (~170 lines of forwarded
  trait impls and never-called with_* builders) with a type alias to
  reth's RpcAddOns, which already provides every forwarded impl.
- Fold crates/network and crates/node-builder into a new
  alethia-reth-node::components module; the three zero-sized component
  builders and ProviderTaikoBlockReader move next to their only
  consumer, TaikoNode::components_builder(). 14 crates -> 12.
- Drop the unused EthereumEthApiBuilder field from TaikoEthApiBuilder
  and the bare TaikoEthApi re-alias of reth's EthApi.

No behavior change; all 191 tests pass.
