+++ Cargo.toml
@@ -7,9 +7,7 @@ members = [
     "crates/consensus",
     "crates/db",
     "crates/evm",
-    "crates/network",
     "crates/node",
-    "crates/node-builder",
     "crates/payload",
     "crates/primitives",
     "crates/rpc",
