+++ crates/node/Cargo.toml
@@ -21,8 +21,6 @@ alethia-reth-chainspec = { path = "../chainspec" }
 alethia-reth-consensus = { path = "../consensus" }
 alethia-reth-db = { path = "../db" }
 alethia-reth-evm = { path = "../evm" }
-alethia-reth-network = { path = "../network" }
-alethia-reth-node-builder = { path = "../node-builder" }
 alethia-reth-payload = { path = "../payload" }
 alethia-reth-primitives = { path = "../primitives", features = ["serde"] }
 alethia-reth-rpc = { path = "../rpc" }
@@ -45,6 +43,7 @@ reth-node-builder = { workspace = true }
 reth-node-ethereum = { workspace = true }
 reth-optimism-trie = { workspace = true, features = ["metrics"] }
 reth-primitives-traits = { workspace = true }
+reth-provider = { workspace = true }
 reth-rpc = { workspace = true }
 reth-rpc-builder = { workspace = true }
 reth-rpc-eth-api = { workspace = true }
