+++ crates/rpc/Cargo.toml
@@ -50,7 +50,6 @@ reth-evm-ethereum = { workspace = true }
 reth-node-api = { workspace = true }
 reth-node-builder = { workspace = true }
 reth-node-core = { workspace = true }
-reth-node-ethereum = { workspace = true }
 reth-optimism-trie = { workspace = true }
 reth-payload-primitives = { workspace = true }
 reth-primitives-traits = { workspace = true }
