+++ crates/rpc/src/eth/builder.rs
@@ -2,16 +2,14 @@ use alethia_reth_block::config::TaikoEvmConfig;
 use alethia_reth_chainspec::spec::TaikoChainSpec;
 use alethia_reth_primitives::engine::TaikoEngineTypes;
 
-use crate::eth::types::TaikoEthApi;
 use reth_ethereum::EthPrimitives;
 use reth_node_api::{FullNodeComponents, NodeTypes};
 use reth_node_builder::rpc::{EthApiBuilder, EthApiCtx};
-use reth_node_ethereum::EthereumEthApiBuilder;
-use reth_rpc::eth::core::EthRpcConverterFor;
+use reth_rpc::{EthApi, eth::core::EthRpcConverterFor};
 
-/// Builds [`TaikoEthApi`] for the Taiko node.
+/// Builds the Taiko `eth` API ([`EthApi`]) for the Taiko node.
 #[derive(Debug, Default)]
-pub struct TaikoEthApiBuilder(EthereumEthApiBuilder);
+pub struct TaikoEthApiBuilder;
 
 impl<N> EthApiBuilder<N> for TaikoEthApiBuilder
 where
@@ -23,9 +21,9 @@ where
         >,
 {
     /// The Ethapi implementation this builder will build.
-    type EthApi = TaikoEthApi<N, EthRpcConverterFor<N>>;
+    type EthApi = EthApi<N, EthRpcConverterFor<N>>;
 
-    /// Builds the [`TaikoEthApi`] from the given context.
+    /// Builds the [`EthApi`] from the given context.
     async fn build_eth_api(self, ctx: EthApiCtx<'_, N>) -> eyre::Result<Self::EthApi> {
         Ok(ctx.eth_api_builder().build())
     }
