--- crates/rpc/src/eth/types.rs
@@ -1,4 +1,0 @@
-use reth_rpc::EthApi;
-
-/// `Eth` API implementation for Taiko network.
-pub type TaikoEthApi<N, Rpc> = EthApi<N, Rpc>;
